Fix autodiscovery of Nvidia drivers
This commit switches to installing the headless Nvidia driver package -- not
strictly necessary but it means we avoid pulling in the whole of Xorg.

In addition, we install ocl-icd-opencl-dev, which installs the dynamic ICD
loader which then chainloads the Nvidia openCL and CUDA libraries.

For more on how this works, see https://github.com/KhronosGroup/OpenCL-ICD-Loader